fix: scale the stream deadline by effort and flag a truncated generation
The 300 second deadline was fixed on both sides and reqwest bounds the whole
request, connect through the last streamed byte, so a generation that
legitimately runs longer was killed mid stream. High effort on a large cloud
model with a full resume and job ad sits squarely in that range, which is the
likely throw behind a failure that then vanished into the silent reset. Both
sides now scale by the request's effort, and the renderer stays the outer
bound so the backend's actionable provider error fires first rather than a
generic timeout.

The tables cannot be shared across the IPC boundary, so a test hardcodes the
backend schedule and asserts the renderer exceeds it at every tier. The tier
order is the vendors', where max sits above xhigh, which reads wrong at a
glance and was inverted here at first: max got a shorter deadline than xhigh,
so the highest effort runs got the tightest bound, which is the exact failure
this exists to prevent. The monotonicity tests listed the tiers in the same
wrong order, so they passed against it; only the value assertions caught it.
Both orders are now pinned with a comment saying why.

finish also consulted the stop reason only when the answer was empty, so a
stream that produced text and then reported length took the success path and a
truncated resume was saved and exported as complete. It still completes and
still keeps the partial text, since that text is worth having, but it now
raises a notification through the existing centre rather than presenting a cut
off document as finished.
