+++ apps/desktop/src/renderer/lib/generate/stream-promise.test.ts
@@ -18,7 +18,7 @@
 import { describe, expect, it, vi } from 'vitest';
 
 import type { AppClient } from '../app-client';
-import { awaitAiStream } from './stream-promise';
+import { awaitAiStream, computeStreamTimeoutMs } from './stream-promise';
 
 interface StreamChunk {
   jobId: string;
@@ -154,5 +154,54 @@ describe('awaitAiStream — empty completion rejects (both resolve paths)', () =
     const promise = awaitAiStream(api, 'job-empty-poll', { pollIntervalMs: 1 });
 
     await expect(promise).rejects.toThrow('Generation produced no content. Please try again.');
+  });
+});
+
+describe('computeStreamTimeoutMs — effort scaling', () => {
+  const BASELINE_MS = 5 * 60 * 1000 + 30_000; // STREAM_TIMEOUT_MS + OUTER_BOUND_MARGIN_MS
+
+  it('uses the flat baseline (+ margin) for no, low-tier, or unrecognized effort', () => {
+    for (const effort of [undefined, 'minimal', 'low', 'bogus-provider-string']) {
+      expect(computeStreamTimeoutMs(effort)).toBe(BASELINE_MS);
+    }
+  });
+
+  it('scales up, strictly non-decreasing, by effort tier', () => {
+    // Vendors' ascending tier order: `max` is the TOP tier, above `xhigh`.
+    const tiers = [undefined, 'minimal', 'low', 'medium', 'high', 'xhigh', 'max'];
+    let prev = 0;
+    for (const effort of tiers) {
+      const ms = computeStreamTimeoutMs(effort);
+      expect(ms).toBeGreaterThanOrEqual(prev);
+      prev = ms;
+    }
+    // And the top tier must actually be LARGER than the baseline, not just
+    // equal — otherwise "scales with effort" would be vacuously true.
+    expect(computeStreamTimeoutMs('max')).toBeGreaterThan(BASELINE_MS);
+  });
+
+  // Cross-language relationship (ADR-style pin, see the task this closed): Rust
+  // and TS can't literally share the effort→multiplier table across the IPC
+  // boundary, so this hardcodes the SAME schedule `timeouts::effort_multiplier`
+  // (apps/desktop/src-tauri/src/commands/ai_provider/timeouts.rs) encodes and
+  // asserts the one invariant that actually matters — the renderer timeout for
+  // a given effort must always exceed the backend's own scaled deadline for
+  // that SAME effort, so the backend (an actionable provider error) fires
+  // first, never the renderer's generic timeout. A change to either table that
+  // breaks this relationship must fail HERE, not surface as a support report.
+  it('stays strictly above the mirrored Rust backend deadline for every known effort level', () => {
+    const RUST_STREAM_BASELINE_MS = 300 * 1000; // timeouts::STREAM
+    const rustEffortMultiplier: Record<string, number> = {
+      medium: 1.5,
+      high: 2.0,
+      xhigh: 2.5,
+      max: 3.0,
+    };
+    for (const effort of [undefined, 'minimal', 'low', 'medium', 'high', 'xhigh', 'max']) {
+      const multiplier = (effort ? rustEffortMultiplier[effort] : undefined) ?? 1;
+      const backendMs = RUST_STREAM_BASELINE_MS * multiplier;
+      const rendererMs = computeStreamTimeoutMs(effort);
+      expect(rendererMs).toBeGreaterThan(backendMs);
+    }
   });
 });
